refactor(host-tools): lift transport-neutral host-tool core out of modes/rpc
Refs #2046
